requirements.txt
# Core
opencv-python          >=4.10
numpy                  >=1.26
tqdm                   >=4.66
scikit-image           >=0.23

# Optional heavy bits (only loaded if their files exist)
ultralytics            # YOLOv8
torch              
torchvision         
torchaudio         

# Graph / ML extras (optional)
networkx               


aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorlog==6.10.1
contourpy==1.3.0
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
docopt==0.6.2
einops==0.8.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
HyperPyYAML==1.2.3
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.6.0
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.2.1
numpy==2.0.2
ollama==0.6.1
omegaconf==2.3.0
opencv-python==4.13.0.92
openpyxl==3.1.5
optuna==4.7.0
packaging==26.0
pandas==2.3.3
pillow==11.3.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.46
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2024.8.30
tokenizers==0.22.2
tomli==2.4.0
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.2.28
ultralytics-thop==2.0.18
urllib3==2.6.3
whisper==1.1.10
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0